TestInfrastructure moved to MutationTestInfrastructure project.
