+++ src/DynamoCoreWpf/DynamoCoreWpf.csproj
@@ -149,8 +149,6 @@
     <Compile Include="UI\VisualConfigurations.cs" />
     <Compile Include="Utilities\LibraryDragAndDrop.cs" />
     <Compile Include="Utilities\OnceDisposable.cs" />
-    <Compile Include="TestInfrastructure\ConnectorMutator.cs" />
-    <Compile Include="TestInfrastructure\MutationTestAttribute.cs" />
     <Compile Include="Utilities\DelegateCommand.cs" />
     <Compile Include="Interfaces\IViewModelView.cs" />
     <Compile Include="Interfaces\IVisualizationManager.cs" />
@@ -162,22 +160,6 @@
     <Compile Include="NodeViewCustomization\NodeViewCustomizationLibrary.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\UsageReportingManager.cs" />
-    <Compile Include="TestInfrastructure\AbstractMutator.cs" />
-    <Compile Include="TestInfrastructure\CodeBlockNodeMutator.cs" />
-    <Compile Include="TestInfrastructure\CopyNodeMutator.cs" />
-    <Compile Include="TestInfrastructure\CustomNodeCompatibilityMutator.cs" />
-    <Compile Include="TestInfrastructure\CustomNodeMutator.cs" />
-    <Compile Include="TestInfrastructure\DeleteNodeMutator.cs" />
-    <Compile Include="TestInfrastructure\DirectoryPathMutator.cs" />
-    <Compile Include="TestInfrastructure\DoubleSliderMutator.cs" />
-    <Compile Include="TestInfrastructure\FilePathMutator.cs" />
-    <Compile Include="TestInfrastructure\IntegerSliderMutator.cs" />
-    <Compile Include="TestInfrastructure\ListMutator.cs" />
-    <Compile Include="TestInfrastructure\MutatorDriver.cs" />
-    <Compile Include="TestInfrastructure\NumberInputMutator.cs" />
-    <Compile Include="TestInfrastructure\NumberRangeMutator.cs" />
-    <Compile Include="TestInfrastructure\NumberSequenceMutator.cs" />
-    <Compile Include="TestInfrastructure\StringInputMutator.cs" />
     <Compile Include="NodeViewCustomization\NodeViewCustomizations\CodeBlockNodeViewCustomization.cs" />
     <Compile Include="Commands\AutomationSettings.cs" />
     <Compile Include="Commands\ConnectorCommands.cs" />
@@ -1211,8 +1193,11 @@
     <Resource Include="UI\Images\Gallery\gallery-button-next-prev-states.png" />
     <Resource Include="UI\Images\profile_normal.png" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="MutationInfastructure\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="AfterBuild">
-        <MakeDir Directories="$(OutputPath)\viewExtensions\" /> 
+    <MakeDir Directories="$(OutputPath)\viewExtensions\" />
   </Target>
 </Project>
