+++ test/MutationTestInfrastructure/ConnectorMutator.cs
@@ -1,20 +1,20 @@
 ﻿using Dynamo.Models;
 using Dynamo.Nodes;
 using Dynamo.Utilities;
-using Dynamo.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.IO;
 using System.Threading;
 using Dynamo.DSEngine;
+using MutationTestInfrastructure;
 
 namespace Dynamo.TestInfrastructure
 {
     [MutationTest("ConnectorMutator")]
     class ConnectorMutator : AbstractMutator
     {
-        public ConnectorMutator(DynamoViewModel viewModel)
+        public ConnectorMutator(DynamoViewModelMutation viewModel)
             : base(viewModel)
         {
         }
